feat: ShortsStateEnum 상태값을 Post 에서 Shorts 엔티티에서 관리
