+++ src/main/java/com/nexters/phochak/domain/Shorts.java
@@ -1,5 +1,6 @@
 package com.nexters.phochak.domain;
 
+import com.nexters.phochak.specification.ShortsStateEnum;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -11,12 +12,16 @@ import javax.persistence.Id;
 
 @Getter
 @Entity
+@Getter
 public class Shorts {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Column(name="SHORTS_ID")
     private Long id;
+
+    @Enumerated(EnumType.STRING)
+    private ShortsStateEnum shortsStateEnum;
 
     @Column(nullable = false, unique = true)
     private String uploadKey;
@@ -31,10 +36,15 @@ public class Shorts {
     }
 
     @Builder
-    public Shorts(Long id, String shortsUrl, String thumbnailUrl) {
+    public Shorts(Long id, String uploadKey, String shortsUrl, String thumbnailUrl) {
         this.id = id;
+        this.shortsStateEnum = ShortsStateEnum.IN_PROGRESS;
         this.uploadKey = uploadKey;
         this.shortsUrl = shortsUrl;
         this.thumbnailUrl = thumbnailUrl;
+    }
+
+    public void updateShortsState(ShortsStateEnum shortsStateEnum) {
+        this.shortsStateEnum = shortsStateEnum;
     }
 }
